sqltelemetry: move record.go to `sql`
`RecordError` requires the import of `log` - but there's no reason it
should be in `sqltelemetry`. Moving that to the `sql` package.

Release note: None
